requirements.txt
blinker==1.4
certifi==2021.10.8
charset-normalizer==2.0.6
click==8.0.2
Flask==2.0.2
Flask-Dance==5.0.0
idna==3.2
itsdangerous==2.0.1
Jinja2==3.0.2
MarkupSafe==2.0.1
oauthlib==3.1.1
python-dotenv==0.19.1
requests==2.26.0
requests-oauthlib==1.3.0
urllib3==1.26.7
URLObject==2.4.3
Werkzeug==2.0.2
